// packages/client/lib/schema-types.ts
export interface CreateCodeReviewRequestMutation_createCodeReviewRequest_errors {
  path: string;
  message: string;
}

export interface CreateCodeReviewRequestMutation_createCodeReviewRequest_codeReviewRequest {
  id: string;
  numDays: number | null;
  codeUrl: string;
  techTags: string[];
  notes: string;
}

export interface CreateCodeReviewRequestMutation_createCodeReviewRequest {
  errors: CreateCodeReviewRequestMutation_createCodeReviewRequest_errors[] | null;
  codeReviewRequest: CreateCodeReviewRequestMutation_createCodeReviewRequest_codeReviewRequest | null;
}

export interface CreateCodeReviewRequestMutation {
  createCodeReviewRequest: CreateCodeReviewRequestMutation_createCodeReviewRequest;
}

export interface CreateCodeReviewRequestMutationVariables {
  input: CreateCodeReviewRequestInput;
}
// ...
export interface CreateCodeReviewRequestInput {
  numDays?: number | null;
  codeUrl: string;
  techTags: string[];
  notes: string;
}